Refactor watchdog module: split into crash_handler module and move crash handling logic there, call it from watchdog.
This is step 1 of a refactor, which I'll complete after I merge the PR, and can write a test of the watchdog's testable code.
* The watchdog would be split into a watchdog directory, exported members in __init__.py
* the cancelable thread stuff would be part of a Cancelable file.
* the crash stats code would live in crash_stats.py
* the crash handler would be in crash_handler.py
* the remainder would get split as necessary, such as init, terminate, etc being in watchdog.py and reexported
