+++ source/watchdog.py
@@ -4,7 +4,6 @@
 # See the file COPYING for more details.
 
 import sys
-import os
 import time
 from time import perf_counter as _timer
 import threading
@@ -12,22 +11,20 @@ from typing import (
 	Any,
 )
 import inspect
-import json
 import ctypes.wintypes
-from dataclasses import (
-	dataclass,
-	asdict,
-)
 import comtypes
 import winBindings.ole32
 import winBindings.dbgHelp
 import winBindings.kernel32
-from winBindings.kernel32 import UnhandledExceptionFilter
 import winUser
 import winKernel
 from logHandler import log
 import logHandler
-import globalVars
+from crash_handler import (
+	CRASH_STATS,
+	crashHandler,
+	loadRecentCrashTimestamps,
+)
 import core
 import exceptions
 import NVDAHelper
@@ -60,52 +57,6 @@ FROZEN_WARNING_TIMEOUT = 15
 """
 
 
-@dataclass(frozen=True)
-class CrashStats:
-	fileName: str = "nvda_crash_stats.txt"
-	timeout: int = 120
-	maxCount: int = 3
-
-	@property
-	def crashStatsPath(self) -> str:
-		return os.path.join(os.path.dirname(globalVars.appArgs.logFileName), self.fileName)
-
-
-_crashStats = CrashStats()
-
-
-@dataclass
-class CrashEvent:
-	timestamp: float
-	version: str
-	installType: str
-
-	def to_json(self) -> str:
-		return json.dumps(asdict(self), separators=(",", ":"))
-
-	@staticmethod
-	def from_line(line: str) -> "CrashEvent | None":
-		if not line:
-			return None
-		try:
-			data = json.loads(line)
-		except json.JSONDecodeError:
-			return None
-		if not isinstance(data, dict):
-			return None
-		try:
-			timestamp = float(data["timestamp"])
-			version = data["version"]
-			installType = data["installType"]
-		except (KeyError, TypeError, ValueError):
-			return None
-		if not isinstance(version, str) or not version:
-			return None
-		if not isinstance(installType, str) or not installType:
-			return None
-		return CrashEvent(timestamp=timestamp, version=version, installType=installType)
-
-
 safeWindowClassSet = {
 	"Internet Explorer_Server",
 	"_WwG",
@@ -120,95 +71,6 @@ _coreDeadTimer = winBindings.kernel32.CreateWaitableTimer(None, True, None)
 _suspended = False
 _watcherThread = None
 _cancelCallEvent = None
-
-
-def _getCurrentCrashFingerprint() -> tuple[str, str]:
-	try:
-		import buildVersion
-
-		version = buildVersion.version
-	except Exception:
-		log.debugWarning("Failed to determine NVDA version for crash stats", exc_info=True)
-		version = "unknown"
-
-	installType = "unknown"
-	try:
-		import config
-	except Exception:
-		log.debugWarning("Failed to import config for crash stats", exc_info=True)
-	else:
-		try:
-			if config.isInstalledCopy():
-				installType = "installed"
-			else:
-				installType = "portable"
-		except Exception:
-			log.debugWarning("Failed to determine install type for crash stats", exc_info=True)
-
-	return version, installType
-
-
-def _buildCrashEvent(timestamp: float, version: str, installType: str) -> CrashEvent:
-	return CrashEvent(timestamp=float(timestamp), version=version, installType=installType)
-
-
-def _writeCrashStats(path: str, events: list[CrashEvent]) -> None:
-	try:
-		with open(path, "w", encoding="utf-8") as f:
-			for event in events:
-				f.write(f"{event.to_json()}\n")
-	except OSError:
-		log.debugWarning("Failed to update crash stats file", exc_info=True)
-
-
-def _loadRecentCrashTimestamps(now: float) -> list[float]:
-	path = _crashStats.crashStatsPath
-	# Check existence explicitly rather than catching exceptions, as this check is far faster than catching an expected exception.
-	if not os.path.exists(path):
-		return []
-	try:
-		with open(path, "r", encoding="utf-8") as f:
-			lines = f.readlines()
-	except FileNotFoundError:
-		return []
-	except OSError:
-		log.debugWarning("Failed to read crash stats file", exc_info=True)
-		return []
-
-	recentCrashes: list[float] = []
-	eventsToRetain: list[CrashEvent] = []
-	needsRewrite = False
-	currentVersion, currentInstallType = _getCurrentCrashFingerprint()
-	for line in lines:
-		event = CrashEvent.from_line(line.strip())
-		if event is None:
-			needsRewrite = True
-			continue
-		timestamp = event.timestamp
-		if now - timestamp <= _crashStats.timeout:
-			eventsToRetain.append(event)
-			if event.version == currentVersion and event.installType == currentInstallType:
-				recentCrashes.append(timestamp)
-		else:
-			# Older entries fall outside of the tracking window.
-			needsRewrite = True
-
-	if needsRewrite:
-		_writeCrashStats(path, eventsToRetain)
-
-	return recentCrashes
-
-
-def _recordCrashTimestamp() -> None:
-	path = _crashStats.crashStatsPath
-	version, installType = _getCurrentCrashFingerprint()
-	try:
-		with open(path, "a", encoding="utf-8") as f:
-			event = _buildCrashEvent(time.time(), version, installType)
-			# Append JSON lines instead of writing a list; NVDA is crashing, keep work minimal
-			f.write(f"{event.to_json()}\n")
-	except OSError:
-		log.debugWarning("Failed to append crash stats", exc_info=True)
 
 
 def alive():
@@ -377,31 +239,6 @@ def _recoverAttempt():
 		pass
 
 
-@UnhandledExceptionFilter
-def _crashHandler(exceptionInfo):
-	threadId = winBindings.kernel32.GetCurrentThreadId()
-	# An exception might have been set for this thread.
-	# Clear it so that it doesn't get raised in this function.
-	ctypes.pythonapi.PyThreadState_SetAsyncExc(threadId, None)
-
-	# Write a minidump.
-	dumpPath = os.path.join(os.path.dirname(globalVars.appArgs.logFileName), "nvda_crash.dmp")
-	if not NVDAHelper.localLib.writeCrashDump(dumpPath, exceptionInfo):
-		log.critical("NVDA crashed! Error writing minidump", exc_info=True)
-	else:
-		log.critical("NVDA crashed! Minidump written to %s" % dumpPath)
-
-	# Log Python stacks for every thread.
-	stacks = logHandler.getFormattedStacksForAllThreads()
-	log.info(f"Listing stacks for Python threads:\n{stacks}")
-
-	_recordCrashTimestamp()
-	log.info("Restarting due to crash")
-	# if NVDA has crashed we cannot rely on the queue handler to start the new NVDA instance
-	core.restartUnsafely()
-	return 1  # EXCEPTION_EXECUTE_HANDLER
-
-
 @ctypes.WINFUNCTYPE(None)
 def _notifySendMessageCancelled():
 	caller = inspect.currentframe().f_back
@@ -424,14 +261,14 @@ def initialize():
 		raise RuntimeError("already running")
 	isRunning = True
 	now = time.time()
-	recentCrashes = _loadRecentCrashTimestamps(now)
-	if len(recentCrashes) >= _crashStats.maxCount:
+	recentCrashes = loadRecentCrashTimestamps(now)
+	if len(recentCrashes) >= CRASH_STATS.maxCount:
 		log.error(
-			f"Crash loop detected ({len(recentCrashes)} crashes in {_crashStats.timeout:.0f} seconds). "
+			f"Crash loop detected ({len(recentCrashes)} crashes in {CRASH_STATS.timeout:.0f} seconds). "
 			"Automatic crash recovery will remain disabled until the loop clears.",
 		)
 	else:
-		winBindings.kernel32.SetUnhandledExceptionFilter(_crashHandler)
+		winBindings.kernel32.SetUnhandledExceptionFilter(crashHandler)
 		# Catch application crashes if the handler is enabled.
 	winBindings.ole32.CoEnableCallCancellation(None)
 	# Cache cancelCallEvent.
